adjust the height of image shown in every coure descrilption and make teh ourServics cart clickable and fix the razorpay realted issue
