+++ client/src/Pages/CourseCart.css
@@ -1,0 +1,225 @@
+/* General Styles */
+.cart-container {
+    max-width: 1200px;
+    margin: 40px auto;
+    padding: 20px;
+    font-family: "Poppins", sans-serif;
+    color: #333;
+    /* background: #f9f9f9; */
+    /* border-radius: 10px; */
+    /* box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1); */
+  }
+  
+  /* Title */
+  .cart-title {
+    font-size: 28px;
+    font-weight: 700;
+    margin-bottom: 10px;
+    color: #222;
+    text-align: left;
+  }
+  
+  .cart-count {
+    font-size: 18px;
+    color: #555;
+  }
+  
+  /* Cart Content Layout */
+  .cart-content {
+    display: flex;
+    flex-direction: column;
+    gap: 25px;
+    margin-top: 20px;
+  }
+  
+  .cart-items {
+    flex: 2;
+    padding: 20px;
+    border-radius: 10px;
+    /* box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1); */
+  }
+  
+  /* Empty Cart */
+  .empty-cart {
+    text-align: center;
+    padding: 50px;
+    color: #777;
+  }
+  
+  .empty-cart .Link {
+    text-decoration: none;
+  }
+  
+  .btn-primary {
+    background-color: #007bff;
+    color: white;
+    padding: 12px 18px;
+    border-radius: 5px;
+    text-transform: uppercase;
+    font-weight: 600;
+    transition: 0.3s ease-in-out;
+    display: inline-block;
+  }
+  
+  .btn-primary:hover {
+    background-color: #0056b3;
+  }
+  
+  /* Cart Item */
+  .cart-item {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 15px;
+    margin-bottom: 10px;
+    border-bottom: 1px solid #ddd;
+    transition: 0.3s ease-in-out;
+    border-radius: 8px;
+    background-color: #f5f5f5;
+  }
+  
+  .cart-item:hover {
+    background: #3333331c;
+  }
+
+  .cart-item-container{
+    display: flex;
+    justify-content: space-between;
+  }
+  
+  .cart-item-image {
+    width: 120px;
+    height: 90px;
+    object-fit: fill;
+    border-radius: 8px;
+    margin-right: 15px;
+  }
+  
+  .cart-item-details {
+    flex: 1;
+  }
+  
+  .cart-item-title {
+    font-size: 18px;
+    font-weight: 600;
+    color: #222;
+  }
+  
+  .cart-item-description {
+    font-size: 14px;
+    color: #383535;
+  }
+  
+  /* Price Details */
+  .cart-item-price {
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+    gap: 30px;
+  }
+  
+  .discount-price {
+    font-size: 20px;
+    font-weight: 700;
+    color: #e63946;
+  }
+  
+  .original-price {
+    font-size: 14px;
+    text-decoration: line-through;
+    color: #777;
+  }
+  
+  /* Remove Button */
+  .btn-remove {
+    background-color: #ff4d4d;
+    color: white;
+    padding: 8px 12px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-weight: 600;
+    transition: 0.3s ease-in-out;
+  }
+  
+  .btn-remove:hover {
+    background-color: #cc0000;
+  }
+  
+  /* Cart Summary */
+  .summry-container{
+    display: flex;
+    align-items: center;
+    justify-content: end;
+    background-color: #f1f1f1;
+
+  }
+  .cart-summary {
+    width: 20%;
+    padding: 25px;
+    border-radius: 10px;
+    text-align: start;
+    align-content: center;
+    /* box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1); */
+  }
+  
+  .summary-title {
+    font-size: 20px;
+    font-weight: 600;
+    color: #222;
+    margin-bottom: 15px;
+  }
+  
+  .total-price {
+    font-size: 24px;
+    font-weight: bold;
+    color: #e63946;
+  }
+  
+  .original-total {
+    font-size: 16px;
+    text-decoration: line-through;
+    color: #777;
+  }
+  
+  .discount-info {
+    font-size: 14px;
+    color: #28a745;
+    font-weight: bold;
+  }
+  
+  /* Buttons */
+  .btn-buy {
+    background-color: #007bff;
+    color: white;
+    padding: 12px 20px;
+    border: none;
+    border-radius: 5px;
+    font-size: 16px;
+    text-transform: uppercase;
+    font-weight: 600;
+    margin-top: 10px;
+    cursor: pointer;
+    transition: 0.3s ease-in-out;
+  }
+  
+  .btn-buy:hover {
+    background-color: #0056b3;
+  }
+  
+  /* Recommended Courses Section */
+  .recommended-section {
+    margin-top: 40px;
+  }
+  
+  .recommended-title {
+    font-size: 22px;
+    font-weight: bold;
+    color: #222;
+  }
+  
+  .recommended-carousel {
+    font-size: 16px;
+    color: #555;
+  }
+
