+++ client/src/Pages/Home.jsx
@@ -1,17 +1,16 @@
 import React from "react";
 import AboutUs from "../components/HomeComponents/AboutUs";
-import OurServices from "../components/HomeComponents/ServicesTwo";
 import PopularCourse from "../components/HomeComponents/PopularCourse";
 import Heilight from "../components/HomeComponents/Heilight";
 import HomeCarousel from "../components/HomeComponents/HomeCarousel";
+import OurServices from "../components/HomeComponents/OurServices";
 const Home = () => {
 
   return (
     <div>
       <HomeCarousel/>
       <AboutUs />
-      {/* <OurServices /> */}
-      <OurServices/>
+      <OurServices/>  
       <PopularCourse />
       {/* <Heilight /> */}
     </div>
