+++ client/src/components/Courses/CourseBanner.css
@@ -2,7 +2,7 @@
 .course-section-container {
   position: relative;
   width: 100vw;
-  height: 70vh;
+  height: 50vh;
 }
 
 .course-section-image {
