+++ client/src/components/HomeComponents/CrousalDuplicate.jsx
