+++ client/src/components/HomeComponents/OurServices.css
@@ -1,107 +1,111 @@
-
-/* OurServices.css */
-.our-services-container {
-    padding: 2rem 1rem;
+.our-services {
     text-align: center;
+    padding: 60px 20px;
+    background-color: #f9f9f9;
   }
   
-  .our-services-title-section {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 1rem;
-    margin-bottom: 2rem;
-    width: 100%;
+  .services-header {
+    max-width: 1200px;
+    margin: 0 auto 40px;
   }
   
-  .lines-container {
+  .our-services-heading {
+    font-size: 2.5rem;
+    font-weight: bold;
+    color: #333;
+    margin-bottom: 10px;
+  }
+  
+  .intro {
+    font-size: 1.2rem;
+    color: #555;
+  }
+  
+  .our-services-grid {
     display: flex;
     flex-direction: column;
-    align-items: center;
+    gap: 20px;
+    max-width: 1000px;
+    height: auto;
+    margin: 0 auto;
+    /* align-items: center; */
+    justify-content: center;
   }
   
-  .line {
-    width: 100px;
-    height: 5px;
-    background-color: #00b0ff;
-    margin-bottom: 5px;
+  /* Service Card Styling */
+  .service-card {
+    background: white;
+    padding: 0px 0px 20px 0px;
+    border-radius: 10px;
+    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease-in-out, box-shadow 0.3s ease-in-out;
   }
   
-  .line.short {
-    width: 10vw; /* Responsive width */
-    min-width: 80px;
+  .service-card:hover {
+    transform: translateY(-5px);
+    box-shadow: 0 8px 20px rgba(0, 0, 0, 0.15);
   }
 
-  .line.long {
-    width: 15vw; /* Responsive width */
-    min-width: 120px;
-  }
-  
-  .our-services-title {
-    font-size: 1.8rem;
-    font-weight: bold;
-  }
-  
-  .services-container {
-    display: flex;
-    gap: 2rem;
-    justify-content: center;
-    flex-wrap: wrap;
-  }
-  
-  .service-card {
-    position: relative;
-    width: 300px;
-    height: 250px;
-    overflow: hidden;
-    border-radius: 8px;
-    cursor: pointer;
-    box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.8);
-    transition: transform 0.3s ease-in-out;
-  }
-  
-  .service-card:hover .overlay {
-    opacity: 1;
-  }
-  
-  .service-card:hover .service-image {
-    transform: scale(1.1);
+  .expanded-card {
+    height: auto;
   }
   
   .service-image {
     width: 100%;
-    height: 100%;
-    object-fit: cover;
-    transition: transform 0.3s ease-in-out;
+    height: 400px;
+    object-fit: fill;
+    border-radius: 8px;
   }
   
-  .overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0, 0, 0, 0.85);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    opacity: 0;
-    transition: opacity 0.3s ease-in-out;
-  }
-  
-  .overlay-text {
-    color: white;
+  .service-title {
     font-size: 1.5rem;
-    font-weight: bold;
-    text-align: center;
-    padding: 1rem;
+    color: #333;
+    margin: 15px 0 10px;
+    text-decoration: none;
+  }
+
+  .service-title:hover{
+    color:#0288d1;
   }
   
-  @media (max-width: 768px) {
-    .our-services-container {
+  .service-description {
+    font-size: 1rem;
+    color: #555;
+    margin: 0px 20px;
+    margin-bottom: 15px;
+    text-align: left;
+  }
+  
+  .learn-more {
+    padding: 10px 20px;
+    font-size: 1.3rem;
+    border: none;
+    background-color: #0288d1;
+    color: white;
+    text-decoration: none;
+    cursor: pointer;
+    transition: background 0.3s ease-in-out;
+    border-radius: 5px;
+  }
+  
+  .learn-more:hover {
+    background-color: #0056b3;
+  }
+  
+  /* Responsive Design */
+  @media (min-width: 768px) {
+    .our-services-grid {
+      display: flex;
       flex-direction: column;
-      align-items: center;
+      justify-content: space-between;
     }
-  }  
-
+  }
+  
+  @media (min-width: 1024px) {
+      .our-services-grid {
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+    }
+  }
