+++ client/src/styles/HomeCarousel.css
@@ -27,7 +27,7 @@
     transform: translateX(-50%);
     text-align: center;
     color: white;
-    background: linear-gradient(to top, rgba(0, 0, 0, 1), rgba(255, 255, 255, 0.1));
+    background: linear-gradient(to top, rgba(0, 0, 0, 1), rgba(255, 255, 255, 0));
     padding: 15px;
     border-bottom-right-radius: 10px;
     border-bottom-left-radius: 10px;
@@ -35,6 +35,8 @@
 
   .home-subtitle{
     color: #0688f3;
+    font-style: bold;
+    text-shadow: -2px 0 8px rgba(0, 0, 0, 0.7);
   }
 
   .home-title{
